+++ core-default-angular/index.js
@@ -4,8 +4,7 @@
 - remove the need to check this.optSubGenerators in EVERY function (i.e. find a way to NOT call this generator AT ALL if sub generator is wrong, but hookFor doesn't seem to be able to be conditionally called based on prompts..?)
 
 @toc
-1. askFor
-2. filesAngular
+1. filesAngular
 
 NOTE: uses Yeoman this.spawnCommand call to run commands (since need to handle Windows/different operating systems and can't use 'exec' since that doesn't show (live) output)
 */
@@ -33,53 +32,6 @@ util.inherits(CoreDefaultAngularGenerator, yeoman.generators.NamedBase);
 
 /**
 @toc 1.
-@method askFor
-*/
-CoreDefaultAngularGenerator.prototype.askFor = function askFor() {
-if(this.optSubGenerators.indexOf('core-default-angular') >-1) {
-
-if(!this.optConfigFile) {		//only prompt if don't have config file
-	var cb = this.async();
-
-	var prompts = [
-	];
-
-	this.prompt(prompts, function (props) {
-		var ii, jj, kk, skip, curName;
-		var skipKeys =[];
-		var toInt =[];
-		for(ii =0; ii<prompts.length; ii++) {
-			curName =prompts[ii].name;
-			skip =false;
-			for(jj =0; jj<skipKeys.length; jj++) {
-				if(curName ==skipKeys[jj]) {
-					skip =true;
-					break;
-				}
-			}
-			if(!skip) {		//copy over
-				//convert to integer (from string) if necessary
-				for(kk =0; kk<toInt.length; kk++) {
-					if(curName ==toInt[kk]) {
-						props[curName] =parseInt(props[curName], 10);
-					}
-				}
-				
-				this.options.props[curName] =this[curName] =props[curName];
-			}
-		}
-		
-		//handle some special ones (the skipKeys from above)
-
-		cb();
-	}.bind(this));
-}
-
-}
-};
-
-/**
-@toc 2.
 @method filesAngular
 */
 CoreDefaultAngularGenerator.prototype.filesAngular = function filesAngular() {