+++ helper-commands/index.js
@@ -3,7 +3,6 @@
 - remove the need to check this.optSubGenerators in EVERY function (i.e. find a way to NOT call this generator AT ALL if sub generator is wrong, but hookFor doesn't seem to be able to be conditionally called based on prompts..?)
 
 @toc
-1. askFor
 2. commandsNpm
 3. commandsBower
 4. commandsGrunt
@@ -29,51 +28,6 @@ var HelperCommandsGenerator = module.exports = function HelperCommandsGenerator(
 };
 
 util.inherits(HelperCommandsGenerator, yeoman.generators.NamedBase);
-
-/**
-@toc 1.
-@method askFor
-*/
-HelperCommandsGenerator.prototype.askFor = function askFor() {
-if(this.optSubGenerators.indexOf('helper-commands') >-1) {
-
-	var cb = this.async();
-
-	var prompts = [
-	];
-
-	this.prompt(prompts, function (props) {
-		var ii, jj, kk, skip, curName;
-		var skipKeys =[];
-		var toInt =[];
-		for(ii =0; ii<prompts.length; ii++) {
-			curName =prompts[ii].name;
-			skip =false;
-			for(jj =0; jj<skipKeys.length; jj++) {
-				if(curName ==skipKeys[jj]) {
-					skip =true;
-					break;
-				}
-			}
-			if(!skip) {		//copy over
-				//convert to integer (from string) if necessary
-				for(kk =0; kk<toInt.length; kk++) {
-					if(curName ==toInt[kk]) {
-						props[curName] =parseInt(props[curName], 10);
-					}
-				}
-				
-				this.options.props[curName] =this[curName] =props[curName];
-			}
-		}
-		
-		//handle some special ones (the skipKeys from above)
-
-		cb();
-	}.bind(this));
-
-}
-};
 
 /**
 @toc 2.